[rename] (component) => (main) 으로 파일 옮김
