+++ src/app/(main)/Banner.tsx
