+++ src/app/(main)/CommunitySection.tsx
