+++ src/app/(main)/NewsData.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { loader } from '@/api/naverAPI';
 import { formatToLocaleDateTimeString } from '@/utils/date';
 import ArrowMore from '@/assets/arrow_back_ios_FILL1_wght400_GRAD0_opsz24.svg';
-import tailwindColors from '../../../../tailwind.config';
+import tailwindColors from '../../../tailwind.config';
 
 import type { NewsType } from '@/types/posts';
 import type { TailwindColors } from '@/types/tailwind';
