+++ src/app/(main)/NewsSection.tsx
