+++ src/app/(main)/RankingSection.tsx
