+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Home from './page.client';
-import NewsData from './(main)/(components)/NewsData';
+import NewsData from './(main)/NewsData';
 
 const page = () => {
   return <Home newsData={<NewsData />} />;
